PBTable visibleColumns Fix
In PBTables visible columns are now used by the frontend to selectively display data. Data from all the columns are sent by the backend (that is specified in serializer). settings for visible columns selections are still saved on backend are updated accordingly.
TODO: Boolean values are not deplayed in the table and connected models are still a problem to display.